+++ plugin.video.ardaktuell/resources/lib/rssaddon/http_status_error.py
@@ -1,0 +1,7 @@
+class HttpStatusError(Exception):
+
+    message = ""
+
+    def __init__(self, msg):
+
+        self.message = msg
